+++ src/mindlessgen/qm/base.py
@@ -20,7 +20,9 @@ class QMMethod(ABC):
         self.verbosity = verbosity
 
     @abstractmethod
-    def optimize(self, molecule: Molecule, verbosity: int = 1) -> Molecule:
+    def optimize(
+        self, molecule: Molecule, max_cycles: int | None = None, verbosity: int = 1
+    ) -> Molecule:
         """
         Define the optimization process.